further simplify code; move FCPDecorator to actioins
